backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
alembic==1.14.1
pydantic-settings==2.7.1
pyjwt[cryptography]==2.10.1
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.20
openai==1.60.2
slowapi==0.1.9
httpx==0.28.1
python-dateutil==2.9.0
pytest==8.3.4
pytest-asyncio==0.25.2
